requirements.txt
allure_python_commons==2.11.1
pytest==7.1.3
python-dotenv==0.21.0
selenium==4.5.0
webdriver_manager==3.8.3
PyMySQL~=1.0.2
SQLAlchemy~=1.4.44
pytest-xdist==2.5.0
cryptography==38.0.3
